+++ synapse-edison/src/main/java/de/otto/synapse/edison/trace/MessageTraceController.java
@@ -5,8 +5,8 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Maps;
 import de.otto.edison.navigation.NavBar;
 import de.otto.synapse.channel.ShardPosition;
+import de.otto.synapse.endpoint.EndpointType;
 import de.otto.synapse.message.Header;
-import de.otto.synapse.messagestore.MessageStore;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.actuate.autoconfigure.ManagementServerProperties;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
@@ -15,6 +15,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.servlet.ModelAndView;
 
+import java.time.LocalDateTime;
+import java.time.ZoneId;
 import java.util.LinkedHashMap;
 
 import static de.otto.edison.navigation.NavBarItem.navBarItem;
@@ -22,27 +24,30 @@ import static java.lang.String.format;
 import static java.util.stream.Collectors.toList;
 
 @Controller
-@ConditionalOnBean(name = "traceMessageStore")
+@ConditionalOnBean(MessageTrace.class)
 public class MessageTraceController {
 
-    private final MessageTraces messageTraces;
+    private final MessageTrace messageTrace;
     private final ObjectMapper objectMapper;
 
     @Autowired
-    MessageTraceController(final MessageTraces messageTraces,
+    MessageTraceController(final MessageTrace messageTrace,
                            final NavBar rightNavBar,
                            final ManagementServerProperties managementServerProperties,
                            final ObjectMapper objectMapper) {
-        this.messageTraces = messageTraces;
+        this.messageTrace = messageTrace;
         this.objectMapper = objectMapper;
-        messageTraces.getSenderChannels().forEach(channelName -> {
+        rightNavBar.register(
+                navBarItem(10, "Message Trace", format("%s/messagetrace", managementServerProperties.getContextPath()))
+        );
+        messageTrace.getSenderChannels().forEach(channelName -> {
             rightNavBar.register(
-                    navBarItem(10, "Sender: " + channelName, format("%s/messagetrace/sender/%s", managementServerProperties.getContextPath(), channelName))
+                    navBarItem(11, "Sender: " + channelName, format("%s/messagetrace/sender/%s", managementServerProperties.getContextPath(), channelName))
             );
         });
-        messageTraces.getReceiverChannels().forEach(channelName -> {
+        messageTrace.getReceiverChannels().forEach(channelName -> {
             rightNavBar.register(
-                    navBarItem(20, "Receiver: " + channelName, format("%s/messagetrace/receiver/%s", managementServerProperties.getContextPath(), channelName))
+                    navBarItem(12, "Receiver: " + channelName, format("%s/messagetrace/receiver/%s", managementServerProperties.getContextPath(), channelName))
             );
         });
     }
@@ -53,22 +58,47 @@ public class MessageTraceController {
     )
     public ModelAndView getMessageTrace(final @PathVariable String endpointType,
                                         final @PathVariable String channelName) {
-        final MessageStore messageStore = endpointType.equals("receiver")
-                ? messageTraces.getReceiverTrace(channelName)
-                : messageTraces.getSenderTrace(channelName);
         return new ModelAndView(
-                "messagetrace",
+                "single-channel-trace",
                 ImmutableMap.of(
                         "title", endpointType.equals("receiver") ? "Receiver: " : "Sender: " + channelName,
-                        "channelName", channelName,
                         "messages",
-                        messageStore
-                                .stream()
-                                .map(message -> ImmutableMap.of(
-                                        "key", message.getKey(),
-                                        "header", prettyPrint(message.getHeader()),
-                                        "payload", prettyPrint(message.getPayload())
+                        messageTrace
+                                .stream(channelName, EndpointType.valueOf(endpointType.toUpperCase()))
+                                .map(traceEntry -> ImmutableMap.of(
+                                        "sequenceNumber", traceEntry.getSequenceNumber(),
+                                        "key", traceEntry.getMessage().getKey(),
+                                        "header", prettyPrint(traceEntry.getMessage().getHeader()),
+                                        "payload", prettyPrint(traceEntry.getMessage().getPayload())
                                 ))
+                                .collect(toList())
+                )
+        );
+    }
+
+    @GetMapping(
+            path = "${management.context-path}/messagetrace",
+            produces = "text/html"
+    )
+    public ModelAndView getMessageTrace() {
+        return new ModelAndView(
+                "multi-channel-trace",
+                ImmutableMap.of(
+                        "title",
+                        "Message Trace",
+                        "messages",
+                        messageTrace
+                                .stream()
+                                .map(traceEntry -> ImmutableMap.builder()
+                                        .put("sequenceNumber", traceEntry.getSequenceNumber())
+                                        .put("ts", LocalDateTime.ofInstant(traceEntry.getTimestamp(), ZoneId.systemDefault()).toString())
+                                        .put("channelName", traceEntry.getChannelName())
+                                        .put("endpointType", traceEntry.getEndpointType().name())
+                                        .put("key", traceEntry.getMessage().getKey())
+                                        .put("header", prettyPrint(traceEntry.getMessage().getHeader()))
+                                        .put("payload", prettyPrint(traceEntry.getMessage().getPayload()))
+                                        .build()
+                                )
                                 .collect(toList())
                 )
         );