+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # Release Notes
 
+## 0.9.0
+* Adds annotation @EnableMessageQueueReceiverEndpoint to auto-configure SQS message-queue receivers
+* Adds annotation @MessageQueueConsumer to auto-configure SQS message-queue listeners
+* Adds message tracing for SQS queues to sender- and receiver-endpoints
+* Adds combined message trace for all sender and receiver channels
+* Introduces synapse-testsupport with separate in-memory configurations for message logs and message queues
+
 ## 0.8.0
 **Breaking Change**: Beans need to be qualified.
 * Register two MessageSenderEndpointFactories: 
   * kinesisMessageSenderEndpointFactory for Kinesis
   * sqsMessageSenderEndpointFactory for SQS
-
 
 ## 0.7.3
 * Fix bug in KinesisShardIterator. Read next record when there is no data near the part of the shard pointed to by the ShardIterator. See https://docs.aws.amazon.com/streams/latest/dev/troubleshooting-consumers.html#getrecords-returns-empty.